Migrate chart view editors to plugin architecture and remove legacy editor files
