+++ src/plugins/web_view_chart/editors/FNAbviewchartareaEditor.js
@@ -1,0 +1,19 @@
+export default function FNAbviewchartareaEditor({ ABViewEditorPlugin }) {
+   return class ABAbviewchartareaEditor extends ABViewEditorPlugin {
+      static getPluginKey() {
+         return this.key;
+      }
+
+      static getPluginType() {
+         return "editor-view";
+      }
+
+      static get key() {
+         return "area";
+      }
+
+      constructor(view, base = "interface_editor_viewchart_area") {
+         super(view, base);
+      }
+   };
+}
