+++ src/plugins/web_view_chart/editors/FNAbviewchartlineEditor.js
@@ -1,0 +1,19 @@
+export default function FNAbviewchartlineEditor({ ABViewEditorPlugin }) {
+   return class ABAbviewchartlineEditor extends ABViewEditorPlugin {
+      static getPluginKey() {
+         return this.key;
+      }
+
+      static getPluginType() {
+         return "editor-view";
+      }
+
+      static get key() {
+         return "line";
+      }
+
+      constructor(view, base = "interface_editor_viewchart_line") {
+         super(view, base);
+      }
+   };
+}
