+++ src/plugins/web_view_chart/editors/FNAbviewchartpieEditor.js
@@ -1,0 +1,19 @@
+export default function FNAbviewchartpieEditor({ ABViewEditorPlugin }) {
+   return class ABAbviewchartpieEditor extends ABViewEditorPlugin {
+      static getPluginKey() {
+         return this.key;
+      }
+
+      static getPluginType() {
+         return "editor-view";
+      }
+
+      static get key() {
+         return "pie";
+      }
+
+      constructor(view, base = "interface_editor_viewchart_pie") {
+         super(view, base);
+      }
+   };
+}
