+++ src/plugins/web_view_chart/properties/FNAbviewchart.js
@@ -1,18 +1,29 @@
-/*
- * ABViewChart
- * A Property manager for our ABViewChart definitions
- */
+import FNAbviewchartareaProperties from "./FNAbviewchartarea.js";
+import FNAbviewchartbarProperties from "./FNAbviewchartbar.js";
+import FNAbviewchartlineProperties from "./FNAbviewchartline.js";
+import FNAbviewchartpieProperties from "./FNAbviewchartpie.js";
 
-import FABViewContainer from "./ABViewContainer";
-
-export default function (AB) {
-   const BASE_ID = "properties_abview_chart";
-
+export default function FNAbviewchartProperties({
+   AB,
+   ABViewPropertiesPlugin,
+}) {
+   const FABViewContainer =
+      require("../../../rootPages/Designer/properties/views/ABViewContainer").default;
    const ABViewContainer = FABViewContainer(AB);
    const uiConfig = AB.Config.uiSettings();
-   const L = ABViewContainer.L();
+   const L = AB.Label();
 
-   class ABViewChartProperty extends ABViewContainer {
+   const BASE_ID = "properties_abview_chart";
+
+   const ABAbviewchartProperties = class ABAbviewchartProperties extends ABViewContainer {
+      static getPluginKey() {
+         return "chart";
+      }
+
+      static getPluginType() {
+         return "properties-view";
+      }
+
       constructor() {
          super(BASE_ID, {
             dataviewID: "",
@@ -190,10 +201,6 @@ export default function (AB) {
       updateCharts() {
          // UPDATE charts when parent properties are changed
          this.CurrentView.refreshData();
-
-         // baseView.views().forEach((e) => {
-         //    e.parent.refreshData();
-         // });
       }
 
       populateDataview() {
@@ -359,7 +366,13 @@ export default function (AB) {
       ViewClass() {
          return super._ViewClass("chart");
       }
-   }
+   };
 
-   return ABViewChartProperty;
+   return [
+      ABAbviewchartProperties,
+      FNAbviewchartareaProperties({ AB, ABViewPropertiesPlugin }),
+      FNAbviewchartbarProperties({ AB, ABViewPropertiesPlugin }),
+      FNAbviewchartlineProperties({ AB, ABViewPropertiesPlugin }),
+      FNAbviewchartpieProperties({ AB, ABViewPropertiesPlugin }),
+   ];
 }
