+++ src/plugins/web_view_chart/properties/FNAbviewchartarea.js
@@ -1,18 +1,21 @@
-/*
- * ABViewChartArea
- * A Property manager for our ABViewChartArea definitions
- */
-
-import FABView from "./ABView";
-
-export default function (AB) {
-   const BASE_ID = "properties_abview_chart_bar";
-
-   const ABView = FABView(AB);
+export default function FNAbviewchartareaProperties({
+   AB,
+   ABViewPropertiesPlugin,
+}) {
    const uiConfig = AB.Config.uiSettings();
-   const L = ABView.L();
+   const L = AB.Label();
 
-   class ABViewChartBarProperty extends ABView {
+   const BASE_ID = "properties_abview_chart_area";
+
+   return class ABAbviewchartareaProperties extends ABViewPropertiesPlugin {
+      static getPluginKey() {
+         return "area";
+      }
+
+      static getPluginType() {
+         return "properties-view";
+      }
+
       constructor() {
          super(BASE_ID, {});
 
@@ -20,24 +23,24 @@ export default function (AB) {
       }
 
       static get key() {
-         return "bar";
+         return "area";
       }
 
       ui() {
          return super.ui([
             {
-               name: "barType",
+               name: "areaType",
                view: "richselect",
                label: L("Chart Type"),
                labelWidth: uiConfig.labelWidthLarge,
                options: [
                   {
-                     id: "bar",
-                     value: L("Vertical"),
+                     id: "area",
+                     value: L("Area"),
                   },
                   {
-                     id: "barH",
-                     value: L("Horizontal"),
+                     id: "stackedArea",
+                     value: L("Stacked Area"),
                   },
                ],
                on: {
@@ -47,38 +50,7 @@ export default function (AB) {
                },
             },
             {
-               name: "barPreset",
-               view: "richselect",
-               label: L("Chart Preset"),
-               labelWidth: uiConfig.labelWidthLarge,
-               options: [
-                  {
-                     id: "column",
-                     value: L("Column"),
-                  },
-                  {
-                     id: "stick",
-                     value: L("Stick"),
-                  },
-                  {
-                     id: "alpha",
-                     value: L("Alpha"),
-                  },
-               ],
-               on: {
-                  onChange: () => {
-                     this.onChange();
-                  },
-               },
-            },
-            // {
-            // 	name: 'chartWidth',
-            // 	view: 'counter',
-            // 	min: 1,
-            // 	label: L('ab.component.chart.bar.chartWidth', '*Width')
-            // },
-            {
-               name: "height",
+               name: "chartHeight",
                view: "counter",
                min: 1,
                label: L("Height"),
@@ -189,9 +161,7 @@ export default function (AB) {
        * NOTE: Must be overwritten by the Child Class
        */
       ViewClass() {
-         return super._ViewClass("bar");
+         return super._ViewClass("area");
       }
-   }
-
-   return ABViewChartBarProperty;
+   };
 }
