+++ src/plugins/web_view_chart/properties/FNAbviewchartpie.js
@@ -1,18 +1,21 @@
-/*
- * ABViewChartline
- * A Property manager for our ABViewChartLine definitions
- */
-
-import FABView from "./ABView";
-
-export default function (AB) {
-   const BASE_ID = "properties_abview_chart_line";
-
-   const ABView = FABView(AB);
+export default function FNAbviewchartpieProperties({
+   AB,
+   ABViewPropertiesPlugin,
+}) {
    const uiConfig = AB.Config.uiSettings();
-   const L = ABView.L();
+   const L = AB.Label();
 
-   class ABViewChartLineProperty extends ABView {
+   const BASE_ID = "properties_abview_chart_pie";
+
+   return class ABAbviewchartpieProperties extends ABViewPropertiesPlugin {
+      static getPluginKey() {
+         return "pie";
+      }
+
+      static getPluginType() {
+         return "properties-view";
+      }
+
       constructor() {
          super(BASE_ID, {});
 
@@ -20,24 +23,28 @@ export default function (AB) {
       }
 
       static get key() {
-         return "line";
+         return "pie";
       }
 
       ui() {
          return super.ui([
             {
-               name: "lineType",
+               name: "pieType",
                view: "richselect",
                label: L("Chart Type"),
                labelWidth: uiConfig.labelWidthLarge,
                options: [
                   {
-                     id: "line",
-                     value: L("Line"),
+                     id: "pie",
+                     value: L("Standard"),
                   },
                   {
-                     id: "spline",
-                     value: L("Spline"),
+                     id: "pie3D",
+                     value: L("Pie3D"),
+                  },
+                  {
+                     id: "donut",
+                     value: L("Donut"),
                   },
                ],
                on: {
@@ -47,38 +54,7 @@ export default function (AB) {
                },
             },
             {
-               name: "linePreset",
-               view: "richselect",
-               label: L("Chart Preset"),
-               labelWidth: uiConfig.labelWidthLarge,
-               options: [
-                  {
-                     id: "plot",
-                     value: L("Plot"),
-                  },
-                  {
-                     id: "diamond",
-                     value: L("Diamond"),
-                  },
-                  {
-                     id: "simple",
-                     value: L("Simple"),
-                  },
-               ],
-               on: {
-                  onChange: () => {
-                     this.onChange();
-                  },
-               },
-            },
-            // {
-            // 	name: 'chartWidth',
-            // 	view: 'counter',
-            // 	min: 1,
-            // 	label: L('ab.component.chart.line.chartWidth', '*Width')
-            // },
-            {
-               name: "chartHeight",
+               name: "height",
                view: "counter",
                min: 1,
                label: L("Height"),
@@ -89,21 +65,11 @@ export default function (AB) {
                },
             },
             {
-               name: "stepValue",
+               name: "innerFontSize",
                view: "counter",
                min: 1,
-               label: L("Step"),
-               on: {
-                  onChange: () => {
-                     this.onChange();
-                  },
-               },
-            },
-            {
-               name: "maxValue",
-               view: "counter",
-               min: 1,
-               label: L("Max Value"),
+               label: L("Inner Font Size"),
+               labelWidth: uiConfig.labelWidthXLarge,
                on: {
                   onChange: () => {
                      this.onChange();
@@ -189,9 +155,7 @@ export default function (AB) {
        * NOTE: Must be overwritten by the Child Class
        */
       ViewClass() {
-         return super._ViewClass("line");
+         return super._ViewClass("pie");
       }
-   }
-
-   return ABViewChartLineProperty;
+   };
 }
