+++ src/rootPages/Designer/editors/EditorManager.js
@@ -13,11 +13,11 @@ export default function (AB) {
    [
       require("./views/_ABViewDefault"),
       // require("./views/ABViewCarousel"),
-      require("./views/ABViewChart"),
-      require("./views/ABViewChartArea"),
-      require("./views/ABViewChartBar"),
-      require("./views/ABViewChartLine"),
-      require("./views/ABViewChartPie"),
+      // require("./views/ABViewChart"),
+      // require("./views/ABViewChartArea"),
+      // require("./views/ABViewChartBar"),
+      // require("./views/ABViewChartLine"),
+      // require("./views/ABViewChartPie"),
       // require("./views/ABViewComment"),
       require("./views/ABViewConditionalContainer"),
       require("./views/ABViewContainer"),
